refactor(router): move NavigationInstruction type to Router
